use one layouting algorithm consistently
